the version of RPM is different on maunaloa so I've added a check to see if the source file exists, and if so then do the final distrib step Bug 600
git-svn-id: http://svn.osafoundation.org/chandler/trunk@4731 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/__hardhat__.py
@@ -198,6 +198,13 @@ def distribute(buildenv):
             os.remove(buildenv['outputdir'] + os.sep + compFile1)
         os.rename(compFile1, buildenv['outputdir'] + os.sep + compFile1)
 
+        # write out the compressed image first - if the installer image doesn't
+        # get built for any reason, then the compressed version will be available
+        if buildenv['version'] == 'release':
+            _outputLine(buildenv['outputdir']+os.sep+"enduser", compFile1)
+        else:
+            _outputLine(buildenv['outputdir']+os.sep+"developer", compFile1)
+
         # The end-user installer
         if installTargetFile:
             installSource = os.path.join(buildenv['root'], installTargetFile)
@@ -206,17 +213,13 @@ def distribute(buildenv):
             if os.path.exists(installTarget):
                 os.remove(installTarget)
 
-            os.rename(installSource, installTarget)
+            if os.path.exists(installSource):
+                os.rename(installSource, installTarget)
             
-            if buildenv['version'] == 'release':
-                _outputLine(buildenv['outputdir'] + os.sep + "enduser", installTargetFile)
-            else:
-                _outputLine(buildenv['outputdir'] + os.sep + "developer", installTargetFile)
-        else:
-            if buildenv['version'] == 'release':
-                _outputLine(buildenv['outputdir']+os.sep+"enduser", compFile1)
-            else:
-                _outputLine(buildenv['outputdir']+os.sep+"developer", compFile1)
+                if buildenv['version'] == 'release':
+                    _outputLine(buildenv['outputdir'] + os.sep + "enduser", installTargetFile)
+                else:
+                    _outputLine(buildenv['outputdir'] + os.sep + "developer", installTargetFile)
 
     # remove the distribution directory, since we have a tarball/zip
     if os.access(distDir, os.F_OK):
